another attempt at resolving np.cross typing issues on windows CI

+++ src/oersted/results.py
@@ -140,6 +140,9 @@ def lorentz_force_density(
 
     assert j_density.shape == b_field.shape
     assert j_density.shape[1] == 3
-    jxb: NDArray[float64] = cross(j_density, b_field)
+
+    # This might introduce a copy, though shouldn't be a big hit to the user's
+    # observed performance
+    jxb: NDArray[float64] = cross(j_density, b_field).astype(float64)
 
     return jxb
